requirements.txt
Flask
torch
torchvision
Pillow
numpy
albumentations
tqdm
moviepy
